@@ -10,16 +10,16 @@
   },
   "dependencies": {
     "axios": "^1.7.9",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-router-dom": "^7.1.5",
     "react-scripts": "^5.0.1",
     "stripe": "^17.6.0",
     "typescript": "^4.1.2"
   },
   "devDependencies": {
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2"
+    "@types/react": "^19.0.8",
+    "@types/react-dom": "^19.0.3"
   },
   "eslintConfig": {
     "extends": [
